Refactor writeTemporaryCredentials to be a utils class
